Minified CodeMX output. Started on some light refactoring to get out of boilerplate-land as was forced by CodeMX.
