+++ Inc/freertos_vars.h
@@ -1,6 +1,10 @@
 /* Define to prevent recursive inclusion -------------------------------------*/
-#ifndef __FREERTOS_H
-#define __FREERTOS_H
+#ifndef __FREERTOS_VARS_H
+#define __FREERTOS_VARS_H
+
+#include "FreeRTOS.h"
+#include "cmsis_os.h"
+
 
 // List of semaphore
 osSemaphoreId sem_usb_irq;
@@ -10,4 +14,4 @@ osThreadId thread_motor_0;
 osThreadId thread_motor_1;
 osThreadId thread_usb_cmd;
 
-#endif /* __FREERTOS_H */
+#endif /* __FREERTOS_VARS_H */
