+++ Odrive.ioc
@@ -370,8 +370,8 @@ ProjectManager.AskForMigrate=true
 ProjectManager.BackupPrevious=false
 ProjectManager.CompilerOptimize=2
 ProjectManager.ComputerToolchain=false
-ProjectManager.CoupleFile=true
-ProjectManager.CustomerFirmwarePackage=
+ProjectManager.CoupleFile=false
+ProjectManager.CustomerFirmwarePackage=/home/sirus/STM32Cube/Repository/STM32Cube_FW_F4_V1.15.0
 ProjectManager.DefaultFWLocation=true
 ProjectManager.DeletePrevious=true
 ProjectManager.DeviceId=STM32F405RGTx
