+++ Src/freertos.c
@@ -44,117 +44,27 @@
 /* Includes ------------------------------------------------------------------*/
 #include "FreeRTOS.h"
 #include "task.h"
-#include "cmsis_os.h"
 
 /* USER CODE BEGIN Includes */     
-#include "freertos_vars.h"
-#include "low_level.h"
-#include "version.h"
+
 /* USER CODE END Includes */
 
 /* Variables -----------------------------------------------------------------*/
-osThreadId defaultTaskHandle;
 
 /* USER CODE BEGIN Variables */
-extern PCD_HandleTypeDef hpcd_USB_OTG_FS;
+
 /* USER CODE END Variables */
 
 /* Function prototypes -------------------------------------------------------*/
-void StartDefaultTask(void const * argument);
-
-extern void MX_USB_DEVICE_Init(void);
-void MX_FREERTOS_Init(void); /* (MISRA C 2004 rule 8.1) */
 
 /* USER CODE BEGIN FunctionPrototypes */
-void usb_cmd_thread(void const * argument);
 
 /* USER CODE END FunctionPrototypes */
 
 /* Hook prototypes */
 
-/* Init FreeRTOS */
-
-void MX_FREERTOS_Init(void) {
-  /* USER CODE BEGIN Init */
-       
-  /* USER CODE END Init */
-
-  /* USER CODE BEGIN RTOS_MUTEX */
-  /* add mutexes, ... */
-  /* USER CODE END RTOS_MUTEX */
-
-  /* USER CODE BEGIN RTOS_SEMAPHORES */
-  // Init usb irq binary semaphore, and start with no tolkens by removing the starting one.
-  osSemaphoreDef(sem_usb_irq);
-  sem_usb_irq = osSemaphoreCreate(osSemaphore(sem_usb_irq), 1);
-  osSemaphoreWait(sem_usb_irq, 0);
-  /* USER CODE END RTOS_SEMAPHORES */
-
-  /* USER CODE BEGIN RTOS_TIMERS */
-  /* start timers, add new ones, ... */
-  /* USER CODE END RTOS_TIMERS */
-
-  /* Create the thread(s) */
-  /* definition and creation of defaultTask */
-  osThreadDef(defaultTask, StartDefaultTask, osPriorityIdle, 0, 256);
-  defaultTaskHandle = osThreadCreate(osThread(defaultTask), NULL);
-
-  /* USER CODE BEGIN RTOS_THREADS */
-
-  /* USER CODE END RTOS_THREADS */
-
-  /* USER CODE BEGIN RTOS_QUEUES */
-  /* add queues, ... */
-  /* USER CODE END RTOS_QUEUES */
-}
-
-/* StartDefaultTask function */
-void StartDefaultTask(void const * argument)
-{
-  /* init code for USB_DEVICE */
-  MX_USB_DEVICE_Init();
-
-  /* USER CODE BEGIN StartDefaultTask */
-
-  // Init motor control
-  init_motor_control();
-
-  // Start motor threads
-  osThreadDef(task_motor_0, motor_thread,   osPriorityHigh+1, 0, 512);
-  osThreadDef(task_motor_1, motor_thread,   osPriorityHigh,   0, 512);
-  thread_motor_0 = osThreadCreate(osThread(task_motor_0), &motors[0]);
-  thread_motor_1 = osThreadCreate(osThread(task_motor_1), &motors[1]);
-
-  // Start USB command handling thread
-  osThreadDef(task_usb_cmd, usb_cmd_thread, osPriorityNormal, 0, 512);
-  thread_usb_cmd = osThreadCreate(osThread(task_usb_cmd), NULL);
-
-  //If we get to here, then the default task is done.
-  vTaskDelete(defaultTaskHandle);
-
-  /* USER CODE END StartDefaultTask */
-}
-
 /* USER CODE BEGIN Application */
-
-// Thread to handle deffered processing of USB interrupt
-void usb_cmd_thread(void const * argument) {
-
-  for (;;) {
-    // Wait for signalling from USB interrupt (OTG_FS_IRQHandler)
-    osSemaphoreWait(sem_usb_irq, osWaitForever);
-    // Irq processing loop
-    //while(HAL_NVIC_GetActive(OTG_FS_IRQn)) {
-      HAL_PCD_IRQHandler(&hpcd_USB_OTG_FS);
-    //}
-    // Let the irq (OTG_FS_IRQHandler) fire again.
-    HAL_NVIC_EnableIRQ(OTG_FS_IRQn);
-  }
-
-  // If we get here, then this task is done
-  vTaskDelete(osThreadGetId());
-}
-
+     
 /* USER CODE END Application */
 
 /************************ (C) COPYRIGHT STMicroelectronics *****END OF FILE****/
